Able to make it into an actual class object and started adding data.
Making more elements will mean adding to the json db so not super quick and
mostly is just a case of adding more functions to the class.

Would also like to format it to a sheet at the end, maybe with a jinja template

Another nice one would be to break the class into a module to practice that.